backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.8.3
httpx==0.27.2
websockets==13.1
pydantic==2.9.2
pydantic-settings==2.5.2
Pillow==10.4.0
qrcode==7.4.2
pg8000==1.31.2
firebase-admin==6.5.0
pytest==8.3.3
httpx==0.27.2
